Few Polish translations improvements
Paste as in paste the text / file was translated as Paste as in a substance. I think it should be the former meaning.
Cut as in cut and paste was translated as a cut on your body.
Translation of FileManager_ButtonYes was not capitalized but translation of FileManager_ButtonNo was, I made it both capitalized
Pan was translated as if it was a Pan for cooking.